requirements.txt
telebot~=0.0.5
python-dotenv~=1.0.1
Django==5.0.6